Color ICP Done, works very well! include a test as Example in apps folder

+++ src/util/fast_eigen_val_vec.cuh
@@ -37,11 +37,11 @@ __forceinline__ __device__ static float3 __compute_eigenvector_1(const float3x3 
 __forceinline__ __device__ static float3 __compute_eigenvector_2(const float3x3 &A,
                                                                  const float3 &evec1, float eval2)
 {
-    float max_evec1_abs = fmaxf(fabsf(evec1.x), fabsf(evec1.y));
-    float inv_length = 1 / sqrtf(max_evec1_abs * max_evec1_abs + evec1.z * evec1.z);
+    float max_evec1_abs = fmax(fabs(evec1.x), fabs(evec1.y));
+    float inv_length = 1.0f / sqrtf(max_evec1_abs * max_evec1_abs + evec1.z * evec1.z);
 
-    float3 U = (fabsf(evec1.x) > fabsf(evec1.y)) ? make_float3(-evec1.z, 0, evec1.x)
-                                                 : make_float3(0, evec1.z, -evec1.y);
+    float3 U = (fabs(evec1.x) > fabs(evec1.y)) ? make_float3(-evec1.z, 0, evec1.x)
+                                               : make_float3(0, evec1.z, -evec1.y);
 
     U *= inv_length;
 
@@ -55,14 +55,14 @@ __forceinline__ __device__ static float3 __compute_eigenvector_2(const float3x3 
     float m01 = dot(U, AV);
     float m11 = dot(V, AV) - eval2;
 
-    float abs_m00 = fabsf(m00);
-    float abs_m01 = fabsf(m01);
-    float abs_m11 = fabsf(m11);
+    float abs_m00 = fabs(m00);
+    float abs_m01 = fabs(m01);
+    float abs_m11 = fabs(m11);
 
     float max_abs_comp;
     if (abs_m00 >= abs_m11)
     {
-        max_abs_comp = fmaxf(abs_m00, abs_m01);
+        max_abs_comp = fmax(abs_m00, abs_m01);
         if (max_abs_comp > 0.0f)
         {
             if (abs_m00 >= abs_m01)
@@ -86,7 +86,7 @@ __forceinline__ __device__ static float3 __compute_eigenvector_2(const float3x3 
     }
     else
     {
-        max_abs_comp = fmaxf(abs_m11, abs_m01);
+        max_abs_comp = fmax(abs_m11, abs_m01);
         if (max_abs_comp > 0.0f)
         {
             if (abs_m11 >= abs_m01)
@@ -119,9 +119,9 @@ __forceinline__ __device__ auto fast_eigen_compute_3x3_symm(const float3x3 &A)
     auto row2 = A.get_row(1);
     auto row3 = A.get_row(2);
 
-    auto abs_max_1 = fmaxf(fmaxf(fabsf(row1.x), fabsf(row1.y)), fabsf(row1.z));
-    auto abs_max_2 = fmaxf(abs_max_1, fmaxf(fabsf(row2.y), fabsf(row2.z)));
-    auto abs_max_coeff = fmaxf(abs_max_2, fabsf(row3.z));
+    auto abs_max_1 = fmax(fmax(fabs(row1.x), fabs(row1.y)), fabs(row1.z));
+    auto abs_max_2 = fmax(abs_max_1, fmax(fabs(row2.y), fabs(row2.z)));
+    auto abs_max_coeff = fmax(abs_max_2, fabs(row3.z));
     if (abs_max_coeff == 0.0f)
     {
         // A is the zero matrix.
@@ -146,27 +146,27 @@ __forceinline__ __device__ auto fast_eigen_compute_3x3_symm(const float3x3 &A)
         float b11 = row2.y - q;
         float b22 = row3.z - q;
 
-        float p = sqrtf((b00 * b00 + b11 * b11 + b22 * b22 + norm * 2) / 6);
+        float p = sqrtf((b00 * b00 + b11 * b11 + b22 * b22 + norm * 2.0f) / 6.0f);
 
         float c00 = b11 * b22 - row2.z * row2.z;
         float c01 = row1.y * b22 - row2.z * row1.z;
         float c02 = row1.y * row2.z - b11 * row1.z;
         float det = (b00 * c00 - row1.y * c01 + row1.z * c02) / (p * p * p);
 
-        float half_det = det * 0.5;
-        half_det = fminf(fmaxf(half_det, -1.0), 1.0);
+        float half_det = det * 0.5f;
+        half_det = fmin(fmax(half_det, -1.0f), 1.0f);
 
         float angle = acos(half_det) / 3.0f;
-        float constexpr two_thirds_pi = 2.09439510239319549;
-        float beta2 = cos(angle) * 2;
-        float beta0 = cos(angle + two_thirds_pi) * 2;
+        float constexpr two_thirds_pi = 2.09439510239319549f;
+        float beta2 = cos(angle) * 2.0f;
+        float beta0 = cos(angle + two_thirds_pi) * 2.0f;
         float beta1 = -(beta0 + beta2);
 
         auto eval_1 = q + p * beta0;
         auto eval_2 = q + p * beta1;
         auto eval_3 = q + p * beta2;
 
-        if (half_det >= 0)
+        if (half_det >= 0.0f)
         {
             float3x3 A_(row1, row2, row3);
             auto evec_3 = __compute_eigenvector_1(A_, eval_3);
